SubsetTreeView was extracted to TreeViewSpinner to be usable at multiple places
